+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Imp/GovSpec.hs
@@ -2,6 +2,7 @@
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE OverloadedLists #-}
 {-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE PatternSynonyms #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
@@ -9,6 +10,12 @@
 
 module Test.Cardano.Ledger.Conway.Imp.GovSpec where
 
+import Cardano.Ledger.Allegra.Scripts (
+  pattern RequireAllOf,
+  pattern RequireAnyOf,
+  pattern RequireMOf,
+  pattern RequireSignature,
+ )
 import Cardano.Ledger.BaseTypes
 import Cardano.Ledger.CertState (vsNumDormantEpochsL)
 import Cardano.Ledger.Coin (Coin (Coin))
@@ -17,11 +24,7 @@ import Cardano.Ledger.Conway.Governance
 import Cardano.Ledger.Conway.Rules (ConwayGovPredFailure (..))
 import Cardano.Ledger.Credential (Credential (KeyHashObj))
 import Cardano.Ledger.DRep (drepExpiryL)
-import Cardano.Ledger.Keys (
-  KeyHash,
-  KeyRole (..),
- )
-import Cardano.Ledger.Plutus (Language (..))
+import Cardano.Ledger.Keys (KeyHash, KeyRole (..))
 import Cardano.Ledger.Shelley.LedgerState
 import Control.Monad (replicateM_)
 import Control.State.Transition.Extended (PredicateFailure)
@@ -32,7 +35,6 @@ import Data.Maybe
 import qualified Data.Sequence.Strict as SSeq
 import qualified Data.Set as Set
 import Lens.Micro
-import Test.Cardano.Ledger.Alonzo.Arbitrary (alwaysSucceeds)
 import Test.Cardano.Ledger.Conway.ImpTest
 import Test.Cardano.Ledger.Core.Rational (IsRatio (..))
 import Test.Cardano.Ledger.Imp.Common
@@ -96,8 +98,14 @@ spec =
           (govActionId, _) <- submitConstitution SNothing
           modifyNES $ \nes ->
             nes
-              & nesEsL . esLStateL . lsUTxOStateL . utxosGovStateL . cgEnactStateL . ensPrevConstitutionL
-                .~ SJust (PrevGovActionId govActionId) -- Add first proposal to PrevGovActionIds in enacted state
+              & nesEsL
+                . esLStateL
+                . lsUTxOStateL
+                . utxosGovStateL
+                . cgEnactStateL
+                . ensPrevConstitutionL
+                .~ SJust (PrevGovActionId govActionId)
+              -- Add first proposal to PrevGovActionIds in enacted state
               & nesEsL . esLStateL . lsUTxOStateL . utxosGovStateL . cgProposalsL
                 .~ def -- Remove all proposals, so that the lookup only succeeds for enacted state.
                 -- Once a proposal with a purpose has been enacted, following proposals can no
@@ -151,12 +159,14 @@ spec =
         impAnn "Constitution has not been enacted yet" $
           curConstitution' `shouldBe` curConstitution
 
-        ConwayGovState expectedProposals expectedEnactState expectedPulser <- getsNES newEpochStateGovStateL
+        ConwayGovState expectedProposals expectedEnactState expectedPulser <-
+          getsNES newEpochStateGovStateL
 
         impAnn "EnactState reflects the submitted governance action" $ do
           let enactStateWithChildren =
                 initialEnactState
-                  & ensPrevGovActionIdsChildrenL . pgacConstitutionL %~ Set.insert (PrevGovActionId govActionId)
+                  & ensPrevGovActionIdsChildrenL . pgacConstitutionL
+                    %~ Set.insert (PrevGovActionId govActionId)
           expectedEnactState `shouldBe` enactStateWithChildren
 
         impAnn "Proposals contain the submitted proposal" $
@@ -175,10 +185,12 @@ spec =
         (dRep, committeeMember) <- electBasicCommittee
         (govActionId, constitution) <- submitConstitution SNothing
 
-        ConwayGovState proposalsBeforeVotes enactStateBeforeVotes pulserBeforeVotes <- getsNES newEpochStateGovStateL
+        ConwayGovState proposalsBeforeVotes enactStateBeforeVotes pulserBeforeVotes <-
+          getsNES newEpochStateGovStateL
         submitYesVote_ (DRepVoter dRep) govActionId
         submitYesVote_ (CommitteeVoter committeeMember) govActionId
-        ConwayGovState proposalsAfterVotes enactStateAfterVotes pulserAfterVotes <- getsNES newEpochStateGovStateL
+        ConwayGovState proposalsAfterVotes enactStateAfterVotes pulserAfterVotes <-
+          getsNES newEpochStateGovStateL
 
         impAnn "Votes are recorded in the proposals" $ do
           let proposalsWithVotes =
@@ -225,10 +237,13 @@ spec =
           enactState <- getsNES $ newEpochStateGovStateL . cgEnactStateL
           rsEnactState pulserRatifyState `shouldBe` enactState
 
-      xit "policy is respected by proposals" $ do
-        let
-          scriptHash = hashScript @era $ alwaysSucceeds @'PlutusV3 0
-          wrongScriptHash = hashScript @era $ alwaysSucceeds @'PlutusV3 1
+      it "policy is respected by proposals" $ do
+        keyHash <- freshKeyHash
+        scriptHash <- impAddNativeScript $ RequireAllOf (SSeq.singleton (RequireSignature keyHash))
+        wrongScriptHash <-
+          impAddNativeScript $
+            RequireMOf 1 $
+              SSeq.fromList [RequireAnyOf mempty, RequireAllOf mempty]
         modifyNES $
           nesEsL . esLStateL . lsUTxOStateL . utxosGovStateL . cgEnactStateL . ensConstitutionL
             .~ Constitution def (SJust scriptHash)
@@ -239,14 +254,13 @@ spec =
               def
                 & ppuCommitteeMinSizeL .~ SJust 1
           rewardAccount <- registerRewardAccount
-          void $
-            submitProposal
-              ProposalProcedure
-                { pProcReturnAddr = rewardAccount
-                , pProcGovAction = ParameterChange SNothing pparamsUpdate (SJust scriptHash)
-                , pProcDeposit = pp ^. ppGovActionDepositL
-                , pProcAnchor = def
-                }
+          submitProposal_
+            ProposalProcedure
+              { pProcReturnAddr = rewardAccount
+              , pProcGovAction = ParameterChange SNothing pparamsUpdate (SJust scriptHash)
+              , pProcDeposit = pp ^. ppGovActionDepositL
+              , pProcAnchor = def
+              }
 
         impAnn "TreasuryWithdrawals with correct policy succeeds" $ do
           rewardAccount <- registerRewardAccount
@@ -255,14 +269,13 @@ spec =
               Map.fromList
                 [ (rewardAccount, Coin 1000)
                 ]
-          void $
-            submitProposal
-              ProposalProcedure
-                { pProcReturnAddr = rewardAccount
-                , pProcGovAction = TreasuryWithdrawals withdrawals (SJust scriptHash)
-                , pProcDeposit = pp ^. ppGovActionDepositL
-                , pProcAnchor = def
-                }
+          submitProposal_
+            ProposalProcedure
+              { pProcReturnAddr = rewardAccount
+              , pProcGovAction = TreasuryWithdrawals withdrawals (SJust scriptHash)
+              , pProcDeposit = pp ^. ppGovActionDepositL
+              , pProcAnchor = def
+              }
 
         impAnn "ParameterChange with invalid policy succeeds" $ do
           rewardAccount <- registerRewardAccount
@@ -278,7 +291,8 @@ spec =
                 , pProcDeposit = pp ^. ppGovActionDepositL
                 , pProcAnchor = def
                 }
-          res `shouldBeLeft` []
+          res
+            `shouldBeLeft` [inject $ InvalidPolicyHash @era (SJust wrongScriptHash) (SJust scriptHash)]
 
         impAnn "TreasuryWithdrawals with invalid policy succeeds" $ do
           rewardAccount <- registerRewardAccount
@@ -295,7 +309,8 @@ spec =
                 , pProcDeposit = pp ^. ppGovActionDepositL
                 , pProcAnchor = def
                 }
-          res `shouldBeLeft` []
+          res
+            `shouldBeLeft` [inject $ InvalidPolicyHash @era (SJust wrongScriptHash) (SJust scriptHash)]
 
     describe "DRep expiry" $ do
       it "is updated based on to number of dormant epochs" $ do
@@ -628,7 +643,7 @@ firstHardForkCantFollow = do
         , pProcAnchor = def
         }
     )
-    [inject @(ConwayGovPredFailure era) (ProposalCantFollow SNothing protver2 protver0)]
+    [inject (ProposalCantFollow @era SNothing protver2 protver0)]
 
 -- | Tests a second hardfork in the Conway era where the PrevGovActionID is SJust
 secondHardForkFollows ::
@@ -648,8 +663,7 @@ secondHardForkCantFollow ::
   forall era.
   ( ShelleyEraImp era
   , ConwayEraTxBody era
-  , -- , GovState era ~ ConwayGovState era
-    Inject (ConwayGovPredFailure era) (PredicateFailure (EraRule "LEDGER" era))
+  , Inject (ConwayGovPredFailure era) (PredicateFailure (EraRule "LEDGER" era))
   ) =>
   ImpTestM era ()
 secondHardForkCantFollow = do
@@ -674,4 +688,4 @@ secondHardForkCantFollow = do
         , pProcAnchor = def
         }
     )
-    [inject @(ConwayGovPredFailure era) (ProposalCantFollow (SJust (PrevGovActionId gaid1)) protver2 protver1)]
+    [inject (ProposalCantFollow @era (SJust (PrevGovActionId gaid1)) protver2 protver1)]