+++ eras/mary/impl/testlib/Test/Cardano/Ledger/Mary/ImpTest.hs
@@ -9,11 +9,12 @@ import Cardano.Crypto.Hash (Hash)
 import Cardano.Ledger.Core (EraIndependentTxBody)
 import Cardano.Ledger.Crypto (Crypto (..))
 import Cardano.Ledger.Mary (MaryEra)
+import Test.Cardano.Ledger.Allegra.ImpTest (impAllegraSatisfyNativeScript)
 import Test.Cardano.Ledger.Common
 import Test.Cardano.Ledger.Mary.TreeDiff ()
 import Test.Cardano.Ledger.Shelley.ImpTest (
   ShelleyEraImp (..),
-  emptyShelleyImpNES,
+  initShelleyImpNES,
  )
 
 instance
@@ -24,4 +25,5 @@ instance
   ) =>
   ShelleyEraImp (MaryEra c)
   where
-  emptyImpNES = emptyShelleyImpNES
+  initImpNES = initShelleyImpNES
+  impSatisfyNativeScript = impAllegraSatisfyNativeScript